Even better TEK support
Actually, you now upgrade firmware on Mac and linux as well, which means
that I can remap the keys in hardware. This is better, because it allows
remapping the keys that would otherwise be indistinguishable in
software.

So I have removed software remapping, and restored ability to switch to
us extended keyboard layout with application button, which is present on
TEK just fine.