Rewrite git-upstream in Elixir
For fun and profit! Also, the new version is smarter. Instead of
requiring user to list remotes it will automatically use all the remotes
defined in the repository if none are provided.

It will also default to upstreaming HEAD if refspec is omitted.

Last, but not least, it knows --usage parameter.