iTerm: inline prompt substitution
Do not use `print` but just inline the substitution. I had trouble with
that, because I did not understand that ${(%)x} first expands variable
x and only *then* applies prompt expansion. To not use a separate
variable, you just prefix the string with :-. That works because empty
variable name is null, and so is replaced with what follows.