Move OSC 7/0 to platform independent part
First of all, I like having directories in tabs in Gnome Terminal, but
this also means that when I SSH from home to work machine, my local
terminal will know the directory (and host name) of my remote session,
instead of potentially a bunch of ~ (ssh) tabs. When I'm too lazy to
tmux, that is.