PS: enable msys2 tig on Windows
I used to just have all of Msys2 in the PATH but now that I'm
experimenting with using Git for Windows, I figured I would go with just
aliasing tig, which is still the best way to add stuff to the index.

Maybe I'll change my mind later and just ensure that Git for Windows is
first in the path, we'll see.